- Add Genesis Transaction Type - Rename Coinbase Transaction Type to 'Staker Reward' Transaction Type - Rework Staker Reward transaction to pay rewards directly into additional stake votes (versus on-chain outputs) - Genesis transaction is only used ONCE at network launch (to initiate funds) - Added additional structure for TransactionHeader to avoid bloading the Staker Reward Transaction - Moved block_index from the transaction to the block information - Reworked transaction fee calculations given changes to Transaction PoW difficulty requirements

+++ src/common/network_fees.h
@@ -12,23 +12,26 @@ namespace TurtleCoin::Common::NetworkFees
 {
     /**
      * Calculate the base transaction fee given the size of the transaction in bytes
+     *
      * @param transaction_size
      * @return
      */
     uint64_t calculate_base_transaction_fee(size_t transaction_size);
 
     /**
-     * Calculate the transaction fee discount given the number of leading zeros
+     * Calculate the transaction fee discount percentage given the number of leading zeros
      * found in the PoW hash of the transaction
+     *
      * @param leading_zeros
      * @return
      */
-    size_t calculate_transaction_discount(size_t leading_zeros = 0);
+    float calculate_transaction_discount(size_t leading_zeros = 0);
 
     /**
      * Calculate the transaction fee given the size of the transaction in bytes
      * as well as the number of leading zeros found in the PoW hash of the
      * transaction
+     *
      * @param transaction_size
      * @param leading_zeros
      * @return
